refactor(jsparser): reduce complexity and break down class into smaller ones
re #14
